requirements.txt
instructor
pydantic
google-generativeai